Started working on the lory(car carry) added a MV(MotorVehicle) count to count amount of vehicles in the world and an MV objects list for ever MV created.

+++ src/MotorVehicle/Ramp.java
@@ -6,5 +6,5 @@ public interface Ramp {
 
     void tiltUp();
     void tiltDown();
-    boolean driveAngel(double angel);
+    boolean driveAngel(int angel);
 }
